Add concern for vote logic
